+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Introduced new API v3 for all players with breaking changes compared to v2
 - Upgraded to Kotlin 1.8.20
 - ExoPlayer: More specific PlaybackException mapping (using errorCodes instead of error types and errorCode names are now part of description)
-- Upgrade example app to latest bitmovin player (v3.39.1), latest ivs player (v1.20.0) and latest exoplayer (v2.19.0)
+- Upgrade example app to latest bitmovin player (v3.40.0), latest ivs player (v1.20.0) and latest exoplayer (v2.19.0)
 - Bitmovin Player: Throw exception if two bitmovin analytics collector instances are attached to the same player instance
 - SourceUrls are autodetected and cannot be specified through metadata anymore